Project Channel Request System
MAJOR overhaul was done to the project channel request system. Project Owners will be able to choose whether they want a channel to public or private, whether they want it to be a text channel or a forum channel, and in what category they want their channel to be in.